// lib/openapi/generated/models/Project.ts
export interface Project {
    /**
     * Unique identifier
     * @type {number}
     * @memberof Project
     */
    readonly id: number;
    /**
     * 
     * @type {ProjectGeneralInformation}
     * @memberof Project
     */
    projectGeneralInformation: ProjectGeneralInformation;
    /**
     * 
     * @type {ProjectSpecificInformation}
     * @memberof Project
     */
    projectSpecificInformation: ProjectSpecificInformation;
    /**
     * 
     * @type {ProjectEnvironmentalImpact}
     * @memberof Project
     */
    projectEnvironmentalImpact: ProjectEnvironmentalImpact;
    /**
     * 
     * @type {ProjectSocialImpact}
     * @memberof Project
     */
    projectSocialImpact: ProjectSocialImpact;
    /**
     * 
     * @type {ProjectCompliance}
     * @memberof Project
     */
    projectCompliance: ProjectCompliance;
}